requirements.txt
sqlalchemy>=2.0.0
faker>=22.0.0
click>=8.1.0
rich>=13.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0
pytest>=7.4.0
httpx>=0.25.0
python-jose[cryptography]>=3.3.0
google-auth>=2.28.0
python-dotenv>=1.0.0
